+++ o2/script.js
@@ -1299,7 +1299,7 @@ function handleYChange(value) {
     const clamped = clampNumber(value, min, max);
     ySlider.value = clamped;
     yInput.value = clamped;
-    setRocketFromCenter(parseInt(clamped));
+    setRocketFromCenter(parseFloat(clamped));
 }
 
 function handleYChangeB(value) {
@@ -1310,7 +1310,7 @@ function handleYChangeB(value) {
     const clamped = clampNumber(value, min, max);
     ySliderB.value = clamped;
     yInputB.value = clamped;
-    setRocketFromCenterB(parseInt(clamped));
+    setRocketFromCenterB(parseFloat(clamped));
 }
 
 function handleSlopeChange(value) {
